remove eea_getTransactionCount and move functionality into prv_get TransactionCount
